Wes is so cool.
