+++ src/main/java/frc/robot/DriverJoystick.java
@@ -2,16 +2,28 @@ package frc.robot;
 
 import frc.robot.constants.DriveConstants;
 import frc.robot.constants.IOConstants;
+import frc.robot.constants.AutoConstants;
+import frc.robot.constants.AutoConstants.DriveWaypoints;
 import frc.robot.libraries.XboxController1038;
 import frc.robot.subsystems.DriveTrain;
 
+import com.pathplanner.lib.auto.AutoBuilder;
+import com.pathplanner.lib.path.GoalEndState;
+import com.pathplanner.lib.path.IdealStartingState;
+import com.pathplanner.lib.path.PathConstraints;
+import com.pathplanner.lib.path.PathPlannerPath;
+
 import edu.wpi.first.math.filter.SlewRateLimiter;
+import edu.wpi.first.math.geometry.Pose2d;
+import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.wpilibj2.command.InstantCommand;
 import edu.wpi.first.wpilibj2.command.RunCommand;
 
 public class DriverJoystick extends XboxController1038 {
     // Subsystem Dependencies
     private final DriveTrain driveTrain = DriveTrain.getInstance();
+
+    private Pose2d currentPose;
 
     // Previous Status
     private double prevX = 0;
@@ -82,6 +94,20 @@ public class DriverJoystick extends XboxController1038 {
 
         // Lock the wheels into an X formation
         super.xButton.whileTrue(this.driveTrain.setX());
+        super.aButton.whileTrue(
+                new InstantCommand(() -> {
+                    this.currentPose = this.driveTrain.getState().Pose;
+                }).andThen(AutoBuilder.followPath(new PathPlannerPath(
+                        PathPlannerPath.waypointsFromPoses(this.currentPose, DriveWaypoints.Algae23.getEndpoint()),
+                        new PathConstraints(
+                                DriveConstants.MaxSpeed,
+                                AutoConstants.kMaxAccelerationMetersPerSecondSquared,
+                                AutoConstants.kMaxAngularSpeedRadiansPerSecond,
+                                AutoConstants.kMaxAngularSpeedRadiansPerSecondSquared),
+                        new IdealStartingState(
+                                driveTrain.getState().Speeds.vxMetersPerSecond,
+                                driveTrain.getState().Pose.getRotation()),
+                        new GoalEndState(0, Rotation2d.kZero)))));
     }
 
     /**
